Upgrade the integration test framework to xunit v3
This upgrades our integration test framework to xunit v3. The migration
is a fairly straightforward migration to the new patterns in xunit v3,
but with one major change on how the actual execution is done. In xunit
v2, many of the test method objects were remotable, which allowed
for the objects to be passed from the test process to tjhe Visual Studio
process. This remoting support has all been removed in xunit v3, so
we need an alternate approach.

The approach taken here is to pass along the assembly name and the list
of test case UniqueIds along which is a list of strings. We rerun the
discovery process in the Visual Studio process, which produces the
same IXunitTestCases we have in the test process. Then we just run
them with any of the existing support. We still create a remotable
wrapper around the message bus (using the built-in JSON serialization
support) and a remoteable wrapper around the discovery/execution
options.